+++ src/App.tsx
@@ -26,7 +26,9 @@ import { PeerConversationChatView } from './Messages/PeerConversationChatView'
 import { BubbleChatView } from './Messages/BubbleChatView';
 import jsonRainbowConfig from './rainbow-config.json';
 import { DialogCallComponent } from './Calls/DialogCallComponent';
-import { ActiveCallBanner } from './Calls/ActiveCallBanner'
+import { ActiveCallBanner } from './Calls/ActiveCallBanner';
+import { SharedFileComponent, FileDescription } from './SharedFile/';
+
 
 const contactsInfoStyle = StyleSheet.create(appStyleConfig.contactsInformation);
 const logger = new Logger('example');
@@ -80,6 +82,8 @@ export default function App() {
                     <Scene key="AddParticipants" component={AddParticipant} />
                     <Scene key="PeerConversationChatView" component={PeerConversationChatView} />
                     <Scene key="BubbleChatView" component={BubbleChatView} />
+                    <Scene key="SharedFiles" component={SharedFileComponent} />
+                    <Scene key="FileDescription" component={FileDescription} />
 
                 </Scene>
             </Router>