+++ src/SharedFile/index.ts
@@ -1,0 +1,2 @@
+export {FileDescription} from './FileDescription'
+export {SharedFileComponent} from './SharedFileComponent'